requirements.txt
asgiref==3.2.10
backports.entry-points-selectable==1.1.1
distlib==0.3.3
Django==3.1.2
django-cors-headers==3.5.0
djangorestframework==3.12.2
Faker==9.8.2
filelock==3.3.2
platformdirs==2.4.0
psycopg2-binary==2.8.6
python-dateutil==2.8.2
pytz==2021.3
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
virtualenv==20.7.2
